Python 3 Migration (#104)
* File number bug fix (#99)

* Updates logic to use the new update_port API

* BX-715 Migration to Python 3

* BX-715 Fixes for unit tests

* BX-715 Updating Docker files for relay and gateway for python 3.7

* BX-715 Fix for transaction stats

* BX-715 Fixes CircleCI build

* BX-715 Docker file fix

* BX-715 Fix CircleCI build

* BX-715 Fixes CircleCI build

* BX-715 Code review fixes

+++ test/unit/messages/eth/test_eth_discovery_message_factory.py
@@ -84,7 +84,7 @@ class EthDiscoveryMessageFactoryTests(AbstractTestCase):
     def _create_half_payload_msg_bytes(self):
         msg_bytes, msg_len = self._create_full_msg_bytes()
         # removing few bytes
-        bytes_to_remove = (len(msg_bytes) - eth_constants.MDC_LEN - eth_constants.SIGNATURE_LEN) / 2
+        bytes_to_remove = int((len(msg_bytes) - eth_constants.MDC_LEN - eth_constants.SIGNATURE_LEN) / 2)
         incomplete_msg_bytes = msg_bytes[:len(msg_bytes) - bytes_to_remove]
 
         return incomplete_msg_bytes, msg_len
@@ -92,7 +92,7 @@ class EthDiscoveryMessageFactoryTests(AbstractTestCase):
     def _create_half_msg_bytes(self):
         msg_bytes, msg_len = self._create_full_msg_bytes()
         # removing few bytes
-        bytes_to_remove = len(msg_bytes) / 2
+        bytes_to_remove = int(len(msg_bytes) / 2)
         incomplete_msg_bytes = msg_bytes[:len(msg_bytes) - bytes_to_remove]
 
         return incomplete_msg_bytes, msg_len
